requirements.txt
telethon==1.20
python-dotenv==0.15.0
orm==0.1.5
aiosqlite==0.17.0
databases==0.4.1
sqlalchemy==1.3.23
isort==5.7.0
black==20.8b1
flake8==3.8.4
ipython==7.21.0

ormar==0.9.8